python/test_github/requirements.txt
requests
bs4
PyGithub
